requirements.txt
django
django-redis
cinemagoer
djangorestframework
django-cors
djangorestframework_simplejwt
pytz